tryna push new stuff
I was editing on the wrong version of this file and i spent a whole hour digging thru my files to find the path to spyder when I could have just got INTO spyder and opened the file there to make the commit thing show up on github. Sigh, what a time.